+++ sdk/src/main/java/com/mapbox/maps/plugin/MapAttributionDelegateImpl.kt
@@ -45,6 +45,7 @@ internal class MapAttributionDelegateImpl constructor(
         .withImproveMap(config.withImproveMap)
         .withTelemetryAttribution(config.withTelemetryAttribution)
         .withMapboxAttribution(config.withMapboxAttribution)
+        .withMapboxPrivacyPolicy(config.withMapboxPrivacyPolicy)
         .withAttributionData(*attributionsArray)
         .build().getAttributions().toList()
     }